Simplify exported list functions
